Move `comms` module to `simulation`
